refactor `master.js` code, allow static creation of new platforms
